requirements.txt
Pillow==8.0.1

Django~=3.1.4